My first attempt
In this attempt I added most if not all things that were asked of. there are some problems with the delete not functioning as i want it to. currently you need to open the side tab/where the menu is to have it update on screen. I don't know why it doesn't update. one other thing i am worried about is that I wasn't allowed to make the side menu collapseble, but i tought it looked better on my phone if it was. I hope they do to.